% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna.R
\name{togo_filter_genes}
\alias{togo_filter_genes}
\title{Remove mito/ribo genes and low-expressed genes from a Seurat object}
\usage{
togo_filter_genes(
  object,
  pct_threshold = 0.5,
  celltype_col = "KPMP_celltype_general"
)
}
\arguments{
\item{object}{Seurat object}

\item{pct_threshold}{Min fraction of cells in ≥1 cell type expressing a gene}

\item{celltype_col}{Metadata column used to stratify expression threshold}
}
\value{
Filtered Seurat object
}
\description{
Remove mito/ribo genes and low-expressed genes from a Seurat object
}
