% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna.R
\name{togo_filter_mito_ribo}
\alias{togo_filter_mito_ribo}
\title{Remove mito/ribo genes and low-expressed genes from a Seurat object}
\usage{
togo_filter_mito_ribo(object, mito_pattern = "^MT-", ribo_genes = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{mito_pattern}{Regex for mitochondrial genes (default human)}

\item{ribo_genes}{Character vector of ribosomal gene names}
}
\value{
Filtered Seurat object
}
\description{
Remove mito/ribo genes and low-expressed genes from a Seurat object
}
