% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna.R
\name{togo_run_seurat_pipeline}
\alias{togo_run_seurat_pipeline}
\title{Run the standard Seurat dim-reduction pipeline}
\usage{
togo_run_seurat_pipeline(object, dims = 1:30, resolution = 0.8)
}
\arguments{
\item{object}{Seurat object}

\item{dims}{PCA dims to use downstream (default 1:30)}

\item{resolution}{Clustering resolution passed to FindClusters}
}
\value{
Processed Seurat object
}
\description{
Run the standard Seurat dim-reduction pipeline
}
